Cleanup of GPU lib, breaking up Resource.h
