+++ libraries/gpu/src/gpu/Format.h
@@ -20,23 +20,6 @@ namespace gpu {
 
 class Backend;
 
-class GPUObject {
-public:
-    virtual ~GPUObject() = default;
-};
-
-class GPUObjectPointer {
-private:
-    using GPUObjectUniquePointer = std::unique_ptr<GPUObject>;
-
-    // This shouldn't be used by anything else than the Backend class with the proper casting.
-    mutable GPUObjectUniquePointer _gpuObject;
-    void setGPUObject(GPUObject* gpuObject) const { _gpuObject.reset(gpuObject); }
-    GPUObject* getGPUObject() const { return _gpuObject.get(); }
-
-    friend class Backend;
-};
-
 // Description of a scalar type
 enum Type {
